[fix] #90 api 수정 및 댓글 paging시에 total page 계산 수정 및 삭제 수정

+++ src/main/java/com/example/umc3_teamproject/service/InterviewParagraphService.java
@@ -99,7 +99,9 @@ public class InterviewParagraphService {
         );
         interviewParagraph.deleted_status_true();
         InterviewParagraphResult interviewParagraphResultByInterviewParagraphId = interviewParagraphResultRepository.findInterviewParagraphResultByInterviewParagraphId(paragraph_id);
-        interviewParagraphResultRepository.delete(interviewParagraphResultByInterviewParagraphId);
+        if(interviewParagraphResultByInterviewParagraphId != null){
+            interviewParagraphResultRepository.delete(interviewParagraphResultByInterviewParagraphId);
+        }
         return new ResponseTemplate<>(paragraph_id+" deleted success");
     }
